Many updates...
Changed old AddWaitingToClientCommand to AddJoinableGameCommand. Changed
old AddJoinableGameCommand to AddGameToJoinableListCommand. This was to
erase any old code of "waiting games." Did the same changes on the
server side.
PlayerCardHand is now accessed through the ClientModel instead of the
Player class.
The GUI should be exactly the same as our Phase 2 pass off code but
connected to the ClientModel.
You can now tell you who are in the game lobby screen. Can't start a
game properly when communicating through the server yet.