GUI updates and PlayerCardHand
Added a missing train card color and removed a non-existent one
The game lobby and waiting lobby now automatically update when they
receive a AddGameToJoinableList and AddPlayerToClient commands.
There's a toast if the server is down, and the Logging in! toast only
appears if you successfully login.
Logout kinda works, but getting the view to reset properly isn't
working. Logout never worked properly in the first place