Fixed Kansas City; added bool to DestCards
Kansas City to St. Louis is a double route but didn't show up since the
mapping name for the blue route was the same as the green.
The boolean in DestCards is to determine on the server side if the
destination path was completed or not. Boolean is on server side too